Create README.md
This project is an approach that uses CPPN-NEAT to evolve weapons and their projectiles they shoot to avoid the obstacles and hit the targets that exist in a given testbed. The core of this tool is based on UnityNEAT, that was created by Daniel Jallov as part of his master's thesis at the Center for Computer Games Research at the IT University in Copenhagen.